WIP chain command
Set the global default on the working directory and started working on
the chain sub-command to scan the directory. Also started writing some
tests and copied over all the "test-chains" from another project to use
for functional testing (since the "scan" and "parse" code works so much
around the filesystem, and I want good coverage).